requirements/pip3requirements.txt
attrdict==2.0.0
certifi==2018.1.18
chardet==3.0.4
cytoolz==0.9.0
ecdsa==0.13
eth-abi==1.0.0
eth-account==0.1.0a2
eth-hash==0.1.0
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.0
eth-tester==0.1.0b11
eth-utils==1.0.1
hexbytes==0.1.0
idna==2.6
lru-dict==1.1.6
py-solc==2.1.0
pycryptodome==3.5.1
pysha3==1.0.2
requests==2.18.4
rlp==0.6.0
semantic-version==2.6.0
six==1.11.0
toolz==0.9.0
urllib3==1.22
web3==4.0.0b11
